+++ src/components/blocks/datePickers/DateSelection/GlobalDateRangeSelection.stories.tsx
@@ -1,6 +1,6 @@
 import { type Meta, type StoryObj } from '@storybook/react-vite'
 
-import { GlobalMonthPicker, type GlobalMonthPickerProps } from '@components/GlobalMonthPicker/GlobalMonthPicker'
+import { GlobalDateRangeSelection, type GlobalDateRangeSelectionProps } from '@blocks/datePickers/DateSelection/GlobalDateRangeSelection'
 
 import { get as getBusiness } from '@msw/api/businesses/[business-id]/get'
 import { handlers } from '@msw/handlers'
@@ -8,40 +8,40 @@ import { makeBusiness } from '@fixtures/business/mocks'
 import { FIXTURE_YEAR, FIXTURE_YEAR_RANGE } from '@fixtures/constants/fixtureYear'
 import { PinnedGlobalDateRange } from '@test-utils/PinnedGlobalDateRange'
 
-const meta: Meta<GlobalMonthPickerProps> = {
-  title: 'Date/GlobalMonthPicker',
-  component: GlobalMonthPicker,
+const meta: Meta<GlobalDateRangeSelectionProps> = {
+  title: 'Date/GlobalDateRangeSelection',
+  component: GlobalDateRangeSelection,
   parameters: {
     msw: { handlers: [getBusiness.mock(makeBusiness({ activationAt: new Date(FIXTURE_YEAR - 1, 0, 1) })), ...handlers] },
-    controls: { include: ['showLabel', 'truncateMonth'] },
+    controls: { include: ['showLabels', 'isCompact'] },
   },
   decorators: [
     Story => (
       <PinnedGlobalDateRange dateRange={FIXTURE_YEAR_RANGE}>
-        <div style={{ padding: '2rem', maxInlineSize: '20rem' }}>
+        <div style={{ padding: '2rem' }}>
           <Story />
         </div>
       </PinnedGlobalDateRange>
     ),
   ],
   args: {
-    showLabel: false,
-    truncateMonth: false,
+    showLabels: false,
+    isCompact: false,
   },
   argTypes: {
-    showLabel: {
+    showLabels: {
       control: 'boolean',
-      description: 'Render the "Month" label above the picker',
+      description: 'Render labels above the range combobox and date pickers',
     },
-    truncateMonth: {
+    isCompact: {
       control: 'boolean',
-      description: 'Use the short month format (e.g. Sep) instead of the full name',
+      description: 'Use the compact layout variant',
     },
   },
 }
 
 export default meta
 
-type Story = StoryObj<GlobalMonthPickerProps>
+type Story = StoryObj<GlobalDateRangeSelectionProps>
 
 export const Default: Story = {}
