+++ src/components/blocks/datePickers/DateSelection/dateGroupByComboBox.scss
