+++ src/components/blocks/datePickers/GlobalDatePicker/GlobalDatePicker.tsx
@@ -3,8 +3,8 @@ import { useTranslation } from 'react-i18next'
 
 import { useBusinessDatePickerBounds } from '@hooks/utils/dates/useBusinessDatePickerBounds'
 import { useGlobalDate, useGlobalDateActions } from '@providers/DateStoreProvider/GlobalDateStoreProvider'
-import { DatePicker } from '@components/DatePicker/DatePicker'
-import { useDatePickerState } from '@components/DatePicker/useDatePickerState'
+import { DatePicker } from '@ui/datePickers/DatePicker/DatePicker'
+import { useDatePickerState } from '@ui/datePickers/DatePicker/useDatePickerState'
 
 export const GlobalDatePicker = ({ showLabel = false }: { showLabel?: boolean }) => {
   const { t } = useTranslation()
