+++ src/components/blocks/datePickers/GlobalDateRangePicker/GlobalDateRangePicker.tsx
@@ -1,5 +1,5 @@
 import { useGlobalDateRange, useGlobalDateRangeActions } from '@providers/DateStoreProvider/GlobalDateStoreProvider'
-import { DateRangePicker } from '@components/DatePicker/DateRangePicker'
+import { DateRangePicker } from '@ui/datePickers/DatePicker/DateRangePicker'
 
 export const GlobalDateRangePicker = ({ showLabels = false }: { showLabels?: boolean }) => {
   const dateRange = useGlobalDateRange({ dateSelectionMode: 'full' })
