+++ src/components/blocks/datePickers/GlobalMonthPicker/GlobalMonthPicker.stories.tsx
@@ -1,6 +1,6 @@
 import { type Meta, type StoryObj } from '@storybook/react-vite'
 
-import { GlobalDateRangeSelection, type GlobalDateRangeSelectionProps } from '@components/DateSelection/GlobalDateRangeSelection'
+import { GlobalMonthPicker, type GlobalMonthPickerProps } from '@blocks/datePickers/GlobalMonthPicker/GlobalMonthPicker'
 
 import { get as getBusiness } from '@msw/api/businesses/[business-id]/get'
 import { handlers } from '@msw/handlers'
@@ -8,40 +8,40 @@ import { makeBusiness } from '@fixtures/business/mocks'
 import { FIXTURE_YEAR, FIXTURE_YEAR_RANGE } from '@fixtures/constants/fixtureYear'
 import { PinnedGlobalDateRange } from '@test-utils/PinnedGlobalDateRange'
 
-const meta: Meta<GlobalDateRangeSelectionProps> = {
-  title: 'Date/GlobalDateRangeSelection',
-  component: GlobalDateRangeSelection,
+const meta: Meta<GlobalMonthPickerProps> = {
+  title: 'Date/GlobalMonthPicker',
+  component: GlobalMonthPicker,
   parameters: {
     msw: { handlers: [getBusiness.mock(makeBusiness({ activationAt: new Date(FIXTURE_YEAR - 1, 0, 1) })), ...handlers] },
-    controls: { include: ['showLabels', 'isCompact'] },
+    controls: { include: ['showLabel', 'truncateMonth'] },
   },
   decorators: [
     Story => (
       <PinnedGlobalDateRange dateRange={FIXTURE_YEAR_RANGE}>
-        <div style={{ padding: '2rem' }}>
+        <div style={{ padding: '2rem', maxInlineSize: '20rem' }}>
           <Story />
         </div>
       </PinnedGlobalDateRange>
     ),
   ],
   args: {
-    showLabels: false,
-    isCompact: false,
+    showLabel: false,
+    truncateMonth: false,
   },
   argTypes: {
-    showLabels: {
+    showLabel: {
       control: 'boolean',
-      description: 'Render labels above the range combobox and date pickers',
+      description: 'Render the "Month" label above the picker',
     },
-    isCompact: {
+    truncateMonth: {
       control: 'boolean',
-      description: 'Use the compact layout variant',
+      description: 'Use the short month format (e.g. Sep) instead of the full name',
     },
   },
 }
 
 export default meta
 
-type Story = StoryObj<GlobalDateRangeSelectionProps>
+type Story = StoryObj<GlobalMonthPickerProps>
 
 export const Default: Story = {}
