+++ src/components/features/generalLedger/ChartOfAccountsTableHeader/ChartOfAccountsTableHeader.tsx
@@ -8,8 +8,8 @@ import { HeaderCol } from '@ui/Header/HeaderCol'
 import { HeaderRow } from '@ui/Header/HeaderRow'
 import { SearchField } from '@ui/SearchField/SearchField'
 import { Heading } from '@ui/Typography/Heading'
+import { LedgerDateRangeSelection } from '@blocks/datePickers/DateSelection/LedgerDateRangeSelection'
 import { ExpandableDataTableToggleButton } from '@blocks/ExpandableDataTable/ExpandableDataTableToggleButton'
-import { LedgerDateRangeSelection } from '@components/DateSelection/LedgerDateRangeSelection'
 import { AccountBalancesDownloadButton } from '@features/generalLedger/AccountBalancesDownloadButton/AccountBalancesDownloadButton'
 import { type ChartOfAccountsTableStringOverrides } from '@features/generalLedger/ChartOfAccountsTableWithPanel/ChartOfAccountsTableWithPanel'
