+++ src/components/features/generalLedger/JournalTableWithPanel/JournalTableWithPanel.tsx
@@ -11,7 +11,7 @@ import { Header } from '@ui/Header/Header'
 import { HeaderCol } from '@ui/Header/HeaderCol'
 import { HeaderRow } from '@ui/Header/HeaderRow'
 import { Heading } from '@ui/Typography/Heading'
-import { LedgerDateRangeSelection } from '@components/DateSelection/LedgerDateRangeSelection'
+import { LedgerDateRangeSelection } from '@blocks/datePickers/DateSelection/LedgerDateRangeSelection'
 import { Panel } from '@components/View/Panel/Panel'
 import { JournalEntriesDownloadButton } from '@features/generalLedger/JournalEntriesDownloadButton/JournalEntriesDownloadButton'
 import { JournalSidebar } from '@features/generalLedger/JournalSidebar/JournalSidebar'
