+++ src/components/features/linkedAccounts/OpeningBalanceAccountForm/OpeningBalanceAccountForm.tsx
@@ -7,12 +7,12 @@ import { type BankAccount } from '@schemas/bankAccounts/bankAccount'
 import { getBankAccountDisplayName, getBankAccountInstitution } from '@utils/bankAccount'
 import { toDataProperties } from '@utils/styleUtils/toDataProperties'
 import { Checkbox } from '@ui/Checkbox/Checkbox'
+import { DatePicker } from '@ui/datePickers/DatePicker/DatePicker'
+import { useDatePickerState } from '@ui/datePickers/DatePicker/useDatePickerState'
 import { AmountInput } from '@ui/Input/AmountInput'
 import { VStack } from '@ui/Stack/Stack'
 import { ErrorText } from '@ui/Typography/ErrorText'
 import { Label, Span } from '@ui/Typography/Text'
-import { DatePicker } from '@components/DatePicker/DatePicker'
-import { useDatePickerState } from '@components/DatePicker/useDatePickerState'
 
 import './openingBalanceAccountForm.scss'
