+++ src/components/features/profitAndLoss/ProfitAndLossReport/ProfitAndLossReport.tsx
@@ -9,7 +9,7 @@ import { Header } from '@ui/Header/Header'
 import { HeaderCol } from '@ui/Header/HeaderCol'
 import { HeaderRow } from '@ui/Header/HeaderRow'
 import { HStack, Stack } from '@ui/Stack/Stack'
-import { CombinedDateRangeSelection } from '@components/DateSelection/CombinedDateRangeSelection'
+import { CombinedDateRangeSelection } from '@blocks/datePickers/DateSelection/CombinedDateRangeSelection'
 import { View } from '@components/View/View'
 import { ProfitAndLossDetailReport } from '@features/profitAndLoss/ProfitAndLossDetailReport/ProfitAndLossDetailReport'
 import { type BreadcrumbItem } from '@features/profitAndLoss/ProfitAndLossDetailReportBreadcrumb/ProfitAndLossDetailReportBreadcrumb'
