+++ src/components/features/reports/BalanceSheet/BalanceSheet.tsx
@@ -11,7 +11,7 @@ import { Header } from '@ui/Header/Header'
 import { HeaderCol } from '@ui/Header/HeaderCol'
 import { HeaderRow } from '@ui/Header/HeaderRow'
 import { HStack, Stack } from '@ui/Stack/Stack'
-import { CombinedDateSelection } from '@components/DateSelection/CombinedDateSelection'
+import { CombinedDateSelection } from '@blocks/datePickers/DateSelection/CombinedDateSelection'
 import { ConditionalBlock } from '@components/utility/ConditionalBlock'
 import { View } from '@components/View/View'
 import { BALANCE_SHEET_ROWS_CONFIG } from '@features/reports/BalanceSheet/constants'
