+++ src/components/features/reports/StatementOfCashFlow/StatementOfCashFlow.tsx
@@ -12,7 +12,7 @@ import { Header } from '@ui/Header/Header'
 import { HeaderCol } from '@ui/Header/HeaderCol'
 import { HeaderRow } from '@ui/Header/HeaderRow'
 import { HStack, Stack } from '@ui/Stack/Stack'
-import { CombinedDateRangeSelection } from '@components/DateSelection/CombinedDateRangeSelection'
+import { CombinedDateRangeSelection } from '@blocks/datePickers/DateSelection/CombinedDateRangeSelection'
 import { ConditionalBlock } from '@components/utility/ConditionalBlock'
 import { View } from '@components/View/View'
 import { CashflowStatementDownloadButton } from '@features/reports/CashflowStatementDownloadButton/CashflowStatementDownloadButton'
