+++ src/components/features/timeTracking/TimeTrackingStats/TimeTrackingStats.tsx
@@ -12,8 +12,8 @@ import { HStack, VStack } from '@ui/Stack/Stack'
 import { DurationSpan } from '@ui/Typography/DurationSpan'
 import { Heading } from '@ui/Typography/Heading'
 import { Span } from '@ui/Typography/Text'
+import { CombinedDateRangeSelection } from '@blocks/datePickers/DateSelection/CombinedDateRangeSelection'
 import { Container } from '@components/Container/Container'
-import { CombinedDateRangeSelection } from '@components/DateSelection/CombinedDateRangeSelection'
 import { ConditionalBlock } from '@components/utility/ConditionalBlock'
 
 import './timeTrackingStats.scss'
