+++ src/components/features/unifiedReports/UnifiedReportControls/UnifiedReportControls.tsx
@@ -11,10 +11,10 @@ import {
   useUnifiedReportReportingBasisParam,
 } from '@providers/UnifiedReportStore/UnifiedReportStoreProvider'
 import { HStack, VStack } from '@ui/Stack/Stack'
-import { CombinedDateRangeSelection } from '@components/DateSelection/CombinedDateRangeSelection'
-import { CombinedDateSelection } from '@components/DateSelection/CombinedDateSelection'
-import { DateGroupByComboBox } from '@components/DateSelection/DateGroupByComboBox'
-import { GlobalYearPicker } from '@components/GlobalYearPicker/GlobalYearPicker'
+import { CombinedDateRangeSelection } from '@blocks/datePickers/DateSelection/CombinedDateRangeSelection'
+import { CombinedDateSelection } from '@blocks/datePickers/DateSelection/CombinedDateSelection'
+import { DateGroupByComboBox } from '@blocks/datePickers/DateSelection/DateGroupByComboBox'
+import { GlobalYearPicker } from '@blocks/datePickers/GlobalYearPicker/GlobalYearPicker'
 import { UnifiedReportHeaderButtons } from '@features/unifiedReports/UnifiedReportHeaderButtons/UnifiedReportHeaderButtons'
 import { UnifiedReportReportingBasisControl } from '@features/unifiedReports/UnifiedReportReportingBasisControl/UnifiedReportReportingBasisControl'
 import { UnifiedReportTagControl } from '@features/unifiedReports/UnifiedReportTagControl/UnifiedReportTagControl'
