+++ src/components/ui/datePickers/MonthCalendar/MonthCalendar.tsx
@@ -9,9 +9,9 @@ import type { View } from '@internal-types/general'
 import { DateFormat } from '@utils/i18n/date/patterns'
 import { useIntlFormatter } from '@hooks/utils/i18n/useIntlFormatter'
 import { Button } from '@ui/Button/Button'
+import { getMonths } from '@ui/datePickers/MonthCalendar/utils'
 import { HStack, VStack } from '@ui/Stack/Stack'
 import { Heading } from '@ui/Typography/Heading'
-import { getMonths } from '@components/MonthCalendar/utils'
 
 import './monthCalendar.scss'
