+++ src/components/ui/datePickers/MonthPicker/MonthPicker.tsx
@@ -6,13 +6,13 @@ import { Dialog, DialogTrigger } from 'react-aria-components/Dialog'
 import { DateFormat } from '@utils/i18n/date/patterns'
 import { useIntlFormatter } from '@hooks/utils/i18n/useIntlFormatter'
 import { useSizeClass } from '@hooks/utils/size/useWindowSize'
+import { MonthCalendar } from '@ui/datePickers/MonthCalendar/MonthCalendar'
 import { Input } from '@ui/Input/Input'
 import { InputGroup } from '@ui/Input/InputGroup'
 import { PickerDropdownIndicator } from '@ui/PickerDropdownIndicator/PickerDropdownIndicator'
 import { VStack } from '@ui/Stack/Stack'
 import { Label } from '@ui/Typography/Text'
 import { ResponsivePopover } from '@blocks/ResponsivePopover/ResponsivePopover'
-import { MonthCalendar } from '@components/MonthCalendar/MonthCalendar'
 
 import './monthPicker.scss'
