+++ src/components/ui/datePickers/MonthPicker/monthPicker.scss
