+++ src/views/MileageTracking.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from 'react-i18next'
 import { Header } from '@ui/Header/Header'
 import { HeaderCol } from '@ui/Header/HeaderCol'
 import { HeaderRow } from '@ui/Header/HeaderRow'
-import { GlobalYearPicker } from '@components/GlobalYearPicker/GlobalYearPicker'
+import { GlobalYearPicker } from '@blocks/datePickers/GlobalYearPicker/GlobalYearPicker'
 import { View } from '@components/View/View'
 import { MileageTrackingStats } from '@features/mileage/MileageTrackingStats/MileageTrackingStats'
 import { TripsRouter } from '@features/mileage/TripsRouter/TripsRouter'
